[x64] Datacache

+++ public/datacache/imdlcache.h
@@ -118,7 +118,7 @@ public:
 	virtual vcollide_t *GetVCollide( MDLHandle_t handle ) = 0;
 	virtual unsigned char *GetAnimBlock( MDLHandle_t handle, int nBlock ) = 0;
 	virtual virtualmodel_t *GetVirtualModel( MDLHandle_t handle ) = 0;
-	virtual int GetAutoplayList( MDLHandle_t handle, unsigned short **pOut ) = 0;
+	virtual intp GetAutoplayList( MDLHandle_t handle, unsigned short **pOut ) = 0;
 	virtual vertexFileHeader_t *GetVertexData( MDLHandle_t handle ) = 0;
 
 	// Brings all data associated with an MDL into memory
@@ -189,7 +189,7 @@ public:
 	virtual vcollide_t *GetVCollide( MDLHandle_t handle ) = 0;
 	virtual unsigned char *GetAnimBlock( MDLHandle_t handle, int nBlock ) = 0;
 	virtual virtualmodel_t *GetVirtualModel( MDLHandle_t handle ) = 0;
-	virtual int GetAutoplayList( MDLHandle_t handle, unsigned short **pOut ) = 0;
+	virtual intp GetAutoplayList( MDLHandle_t handle, unsigned short **pOut ) = 0;
 	virtual vertexFileHeader_t *GetVertexData( MDLHandle_t handle ) = 0;
 
 	// Brings all data associated with an MDL into memory
